looks like we might go with a long in the parser, but this will do for now.

+++ .src/MidiReader.Parse.cs
@@ -35,7 +35,7 @@ namespace on.smfio
     }
     
     /// <inheritdoc/>
-    public string GetMbtString(ulong value) { return TimeUtil.GetMBT(value, Division); }
+    public string GetMbtString(ulong value) { return TimeUtil.GetMBT((long)value, Division); }
 
     //
     // TIME
